refactor: move Trainer/Sensor/ReconstructionError to limen.inference
Trainer and Sensor are inference-time (deployment) concerns, not
experiment-runner concerns. Moving them to limen.inference breaks the
circular import limen.yaml.compiler → limen.experiment.trainer → limen.yaml.compiler.

limen.Trainer, limen.Sensor, limen.ReconstructionError lazy exports preserved.
