+++ limen/__init__.py
@@ -31,10 +31,10 @@ _LAZY_EXPORTS = {
     'Log': ('limen.log.log', 'Log'),
     'MLManifest': ('limen.experiment', 'MLManifest'),
     'Manifest': ('limen.experiment', 'Manifest'),
-    'ReconstructionError': ('limen.experiment', 'ReconstructionError'),
+    'ReconstructionError': ('limen.inference', 'ReconstructionError'),
     'RuleBasedManifest': ('limen.experiment', 'RuleBasedManifest'),
-    'Sensor': ('limen.experiment', 'Sensor'),
-    'Trainer': ('limen.experiment', 'Trainer'),
+    'Sensor': ('limen.inference', 'Sensor'),
+    'Trainer': ('limen.inference', 'Trainer'),
     'UniversalExperimentLoop': ('limen.experiment', 'UniversalExperimentLoop'),
     'features': ('limen.features', None),
     'indicators': ('limen.indicators', None),
