+++ limen/experiment/__init__.py
@@ -6,9 +6,6 @@ from limen.experiment.manifest_core import RuleBasedManifest
 from limen.experiment.param_search import GridStrategy
 from limen.experiment.param_search import RandomStrategy
 from limen.experiment.param_search import STRATEGY_REGISTRY
-from limen.experiment.trainer import ReconstructionError
-from limen.experiment.trainer import Sensor
-from limen.experiment.trainer import Trainer
 
 __all__ = [
     'STRATEGY_REGISTRY',
@@ -17,9 +14,6 @@ __all__ = [
     'MLManifest',
     'Manifest',
     'RandomStrategy',
-    'ReconstructionError',
     'RuleBasedManifest',
-    'Sensor',
-    'Trainer',
     'UniversalExperimentLoop',
 ]
