--- limen/experiment/trainer/__init__.py
@@ -1,9 +1,0 @@
-from limen.experiment.trainer.errors import ReconstructionError
-from limen.experiment.trainer.sensor import Sensor
-from limen.experiment.trainer.trainer import Trainer
-
-__all__ = [
-    'ReconstructionError',
-    'Sensor',
-    'Trainer',
-]
