+++ limen/inference/__init__.py
@@ -1,0 +1,11 @@
+from limen.inference.errors import ReconstructionError
+from limen.inference.sensor import BarPrediction
+from limen.inference.sensor import Sensor
+from limen.inference.trainer import Trainer
+
+__all__ = [
+    'BarPrediction',
+    'ReconstructionError',
+    'Sensor',
+    'Trainer',
+]
